[FILE MOVE] Move the PCS traits out of poly into a new pcs module
